adding call argument matching

+++ testcases/function.c
@@ -1,12 +1,12 @@
-long long foo(long long a, 
-              long long b, 
-              long long c, 
-              long long d, 
-              long long e,
-              long long f,      // up to here should be passed in registers
-              long long g,
-              long long h,
-              long long i)      // these should be stored on stack
+int foo(int a, 
+        int b, 
+        int c, 
+        int d, 
+        int e,
+        int f,      // up to here should be passed in registers
+        int g,
+        int h,
+        int i)      // these should be stored on stack
 {
     return a + b + c + d + e + f + g + h + i;
 }
